requirements.txt
python-dotenv~=1.0.1
envparse~=0.2.0
whisper==1.1.10
python-ffmpeg
openai-whisper
aiohttp~=3.10.8
ffmpeg-python~=0.2.0
aiogram~=3.13.1
scikit-learn~=1.5.2
beanie~=1.27.0
motor~=3.6.0
langchain-text-splitters~=0.3.0
qdrant-client~=1.12.0
FlagEmbedding~=1.2.11
pdfminer.six
peft
pdfminer
pydantic~=2.9.2
openpyxl
python-docx